Fix unit tests
This also fixes a bunch of miscellaneous stuff, giving tools in the
testing suite and fixing a few bugs.

+++ database/migrations/2015_05_23_233151_remove_guide_info_table.php
@@ -14,7 +14,9 @@ class RemoveGuideInfoTable extends Migration
      */
     public function up()
     {
-        Schema::drop('guide_info');
+        if (Schema::hasTable('guide_info')) {
+            Schema::drop('guide_info');
+        }
     }
 
     /**
