+++ database/migrations/2015_05_23_233305_remove_guides_locations_table.php
@@ -14,7 +14,9 @@ class RemoveGuidesLocationsTable extends Migration
      */
     public function up()
     {
-        Schema::drop('guides_locations');
+        if (Schema::hasTable('guides_locations')) {
+            Schema::drop('guides_locations');
+        }
     }
 
     /**
